modify douban movie Yu LIU
